+++ mypy.ini
@@ -141,6 +141,9 @@ disallow_untyped_calls = False
 [mypy-wandb.sdk.data_types]
 disallow_untyped_calls = False
 
+[mypy-wandb.sdk.data_types._dtypes]
+ignore_errors = True
+
 [mypy-wandb.sdk.wandb_artifacts]
 disallow_untyped_calls = False
 
@@ -159,9 +162,6 @@ ignore_errors = True
 # -----------------------------
 # wandb/sdk/interface/
 # -----------------------------
-
-[mypy-wandb.sdk.interface._dtypes]
-ignore_errors = True
 
 [mypy-wandb.sdk.interface.artifacts]
 ignore_errors = True
