+++ wandb/sdk/data_types/Plotly.py
@@ -1,0 +1,89 @@
+from ._WBValue import WBValue
+from ._Media import Media
+from .Image import Image
+
+from ._Media import _numpy_arrays_to_lists
+from ._common import MEDIA_TMP
+
+import codecs
+import os
+from typing import (
+    Sequence,
+    Type,
+    TYPE_CHECKING,
+    Union,
+)
+
+from wandb import util
+
+
+if TYPE_CHECKING:  # pragma: no cover
+    from ..wandb_artifacts import Artifact as LocalArtifact
+    from ..wandb_run import Run as LocalRun
+    import pandas as pd  # type: ignore
+    import matplotlib  # type: ignore
+    import plotly  # type: ignore
+
+    ValToJsonType = Union[
+        dict,
+        "WBValue",
+        Sequence["WBValue"],
+        "plotly.Figure",
+        "matplotlib.artist.Artist",
+        "pd.DataFrame",
+        object,
+    ]
+
+
+class Plotly(Media):
+    """
+    Wandb class for plotly plots.
+
+    Arguments:
+        val: matplotlib or plotly figure
+    """
+
+    _log_type = "plotly-file"
+
+    @classmethod
+    def make_plot_media(
+        cls: Type["Plotly"], val: Union["plotly.Figure", "matplotlib.artist.Artist"]
+    ) -> Union[Image, "Plotly"]:
+        if util.is_matplotlib_typename(util.get_full_typename(val)):
+            if util.matplotlib_contains_images(val):
+                return Image(val)
+            val = util.matplotlib_to_plotly(val)
+        return cls(val)
+
+    def __init__(self, val: Union["plotly.Figure", "matplotlib.artist.Artist"]):
+        super(Plotly, self).__init__()
+        # First, check to see if the incoming `val` object is a plotfly figure
+        if not util.is_plotly_figure_typename(util.get_full_typename(val)):
+            # If it is not, but it is a matplotlib figure, then attempt to convert it to plotly
+            if util.is_matplotlib_typename(util.get_full_typename(val)):
+                if util.matplotlib_contains_images(val):
+                    raise ValueError(
+                        "Plotly does not currently support converting matplotlib figures containing images. \
+                            You can convert the plot to a static image with `wandb.Image(plt)` "
+                    )
+                val = util.matplotlib_to_plotly(val)
+            else:
+                raise ValueError(
+                    "Logged plots must be plotly figures, or matplotlib plots convertible to plotly via mpl_to_plotly"
+                )
+
+        tmp_path = os.path.join(MEDIA_TMP.name, util.generate_id() + ".plotly.json")
+        val = _numpy_arrays_to_lists(val.to_plotly_json())
+        with codecs.open(tmp_path, "w", encoding="utf-8") as fp:
+            util.json_dump_safer(val, fp)
+        self._set_file(tmp_path, is_tmp=True, extension=".plotly.json")
+
+    @classmethod
+    def get_media_subdir(cls: Type["Plotly"]) -> str:
+        return os.path.join("media", "plotly")
+
+    def to_json(self, run_or_artifact: Union["LocalRun", "LocalArtifact"]) -> dict:
+        json_dict = super(Plotly, self).to_json(run_or_artifact)
+        json_dict["_type"] = self._log_type
+        return json_dict
+
